requirements_dev.txt
black
codespell
flake8
pre-commit
isort
pytest>=4.6
pylibiio>=0.21
